Bug #466506: Fix documentation about currency retrieval
This patch corrects the documentation error encountered by the
reporter.  When making this correction, I noticed that the whole section on
disabling retrieval of currency exchange rates was in rather poor shape, so I
went ahead and cleaned up the rest of that section as well and included that
work in this patch.

Patch by Daniel Harding.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash-docs/trunk@17652 57a11ea4-9604-0410-9ed3-97b8803252fd