Spelling, grammar and sentence structure tweaks
A couple of minor spelling corrections and a little tweaking to make some
sentences tighter. Several tweaks to make paragraphs in `Travel Advances'
section gender-neutral, and avoid too many `he or she' phrases. A couple of
small changes to refer to accounts with GnuCash's familiar A:B:C notation,
or with just normal wording. Replace single-digit numbers not being used in
calculations with their word equivalents, as customary in prose.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash-docs/trunk@19628 57a11ea4-9604-0410-9ed3-97b8803252fd