guide: Add What's New section for current stable series
The idea is to always have the What's New section for the current stable
series at the top and move older What's New sections down as they get
superseded.

The list of changes is a greatly summarised one compiled by looking through
the milestone bugfixes pointed to by
http://wiki.gnucash.org/wiki/Release_Schedule to pick up on significant
user-visible improvements.

https://bugzilla.gnome.org/show_bug.cgi?id=628745

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash-docs/trunk@19704 57a11ea4-9604-0410-9ed3-97b8803252fd