Add links for backwards compatibility with older yelp versions
This is a follow up on r22373 which fixes the help file locations
for newer yelp versions. That fix was incompatible with older
yelp versions. With the links in place, both old and new yelp
versions should properly find the help files.

This change is deliberately committed to the 2.4 branch only.
I is expected that the older yelp versions won't be used anymore
by the time the current trunk gets stable.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash-docs/branches/2.4@22471 57a11ea4-9604-0410-9ed3-97b8803252fd