Convert help documents to use XInclude instead of SYSTEM entities
This is mostly the same thing as in the previous commit for guide documents.
Note I have made one additional change in the txf-categories files: they
had two top-level para nodes, but xml only allows one. So I have moved the
content of the second one (which cited the source of the txf-categories)
into the first para to make the builds work again.