Line up xml tag structure in retained files
This is the first step to extract translations into a po file.
When the tag structure across the documents is exactly the same on
the document files across different languages, the extraction
of translatable strings/translated messages can be done in a semi
automated way.
Note some sections exist only in one language. ITS has a mechanism to
handle this, but that can only be used *after* the documents have
been converted to po based translations. So in the interim missing
sections are added with markers that can later be searched for. I am
currently using these markers:
* LANG-DE: will be found in the C documents to indicate this section
  only exists in the German translation, not in the English original
* UNTRANSLATED-DE: used in the German translation to indicate this
  section did not exist in the translation, but does in the English
  original.
* FUZZY-DE: used in the German translation, indicating the context
  of this section has changed and needs review.

Sections marked with LANG-DE or UNTRANSLATED-DE are either candidates
for a special ITS marker (if the section is only relevant for one
language) or for future translation.

Note also that whitespace differences are not important. ITS cleans up
whitespace before extracting the strings. The sequence of the tags in
a file is what matters.