Add pdf targets to CMake build system
This comes with a few additional changes
- Japanese font selection logic is reduced to a minimum. That is,
  one can pass fonts and a font directory on the cmake command line
  with the appropriate -D flags, but none of these are checked for
  correctness. I thought of porting the autotools rules for this
  but looking closer they are totally broken. If really needed
  an improved variation could be written later in cmake.
- Every fop run now requires a fop config file (fop.xconf) in cmake.
  This allows fop to find our figures in the source tree instead
  eliminates the need for a link or copy of these figures in the
  build tree. The existing config files for Russian and Japanese
  are amended for this.
- All settings that just mimic the defaults have been removed from
  the Japanese fop config file.

In addition the target structure has been slightly re-arranged.
There are a few global targets like html, check, pdf, xml.
Each of these will depend on the equivalent targets per document.
For example: the global html target depends on gnucash-guide-html
and gnucash-help-html.
Each of these in turn then depend on the equivalent targets per
language. For example gnucash-help-html in turn depends on
C-gnucash-help-html, it-gnucash-help-html and so on.

This allows to make all documents in one go or only one document
in all supported languages or only one single document.