CMake - add dist and distcheck targets
On a tree for which both configure and cmake have been run, dist will
generate identical tarballs under autotools and cmake, with the exception
of two empty directories.

I haven't tested yet on a clean checkout with only cmake for comparison.
It's likely that will fail because cmake doesn't know how to generate
the many autotools config related files. To be investigated further.