CMake - refactor targets into custom commands
A custom target will always run its commands. However we only want
to run the commands if one of the dependencies changes. This is
exactly how custom commands work. A custom target is still defined as well
for each document/lang/format triplet as that is needed for dependency
management across directories and to have a direct target to pass to make

Note for the html target(s) this posed a challenge. When cmake is run
to set up the build system it is unknown still what the final html files
will be for the html targets so it's impossible to add dependencies at
set up time. Instead a dummy file is created when the html build rule
is run. Whenever any of the dependencies is newer than this dummy file
the build rule will be triggered.