CMake - revisit xml target
This is really a target to install the documentation where gnome help
searches for it rather than some generic xml target. Rename the
target to ghelp to make this more explicit.

In addition add rules to 'install' in the build directory just like
all the other targets. The real install target is also kept as
this is what distro packagers will be using eventually.

Lastly, move the check target to a separate location as it's not
ghelp specific. It is used to run xmllint on the source files
which would benefit all document formats. It has been made optional
though as calling 'make check' will always run the check. Setting
it as dependency for other targets would in turn cause the other
targets to always rebuild even if no source files were changed.