Demo output of xmllint
The first cli xml formatter is already in use by the project: xmllint
But either it is very rudimentary or its man page is not well written.

See the result of 'for i in *.xml; do  xmllint --format --c14n --path
../../docbook/ $i --output output/xmllint/$i; done'

The files still contain [TAB]s and the line length was not limited.