cmake- build chm file by default on Windows
The chm targets were already enabled by default, but not built, that
is one still had to explicitly add the 'chm' target toi make or ninja
for chm files to get built.
This turned out to be a problem for our Windows build scripts
which are driven by jhbuild. While jhbuild allows to set ninjaargs
to drive ninja, that parameter appears to ignore targets and will
only accept command line options.