Fix Cmake dist tarball creation with autotools-build option.
If AUTOTOOLS_IN_DIST was set and the autogen build products (aclocal.m4,
configure, config.guess, etc.) weren't present in CMAKE_SOURCE_ROOT
ninja dist would fail because the dist target was dependent on those
files, even though autogen.sh would create them as part of building the
tarballs.

This was apparently in order to enable building a dist tarball when
autogen.sh wasn't available, but the only reason that might be the case
is if one is starting from a dist tarball. That's silly, so removed
the autogen-built files from dist_files and substituted an error.