Update adjust-dpi.sh so rerun does nothing
Slight rounding issues previously meant some files were updated but
not actually changed. Increase the no of decimals in the existing and
future pixels per centimeter from 1 to 2 and check for a difference of
more than 0.02 pixels per centimeter before running convert on a file.

Included in this commit are images that have not been previously
updated by adjust-dpi.sh.