Fix 2 "[WARN] FOUserAgent - Destination: Unresolved ID reference"
messages when building help pdf. The problem is when linking to an id
defined in a title tag.
1. destination "trans-match-col" in Help_ch_Transactions.xml:
  Html: clicking the link incorrectly takes you to the start of the
  enclosing sect1 tag
  Pdf: link not created
2. destination "report-common.title" in Help_ch_Reports.xml:
  Html: link works
  Pdf: link not created
The fix is to move the id's from the title tag to the previous sectx
tag. This was simple in Help_ch_Reports.xml as there was already an
enclosing sectx tag but in Help_ch_Transactions.xml new sect's had to be
created and this involved much indenting. No content is changed in this
commit.
After the fix, the links work correctly in both html + pdf.