Fix race condition in creating symlinks
For the kde help system, a few extra symlinks are created.
The way this was done however caused the build system to
recreate the exact same symlink from the guide and the manual
build scripts. In many cases this was detected by cmake and
the second attempt skipped, but I ran in to race conditions
where the build would error out while attempting to create
the already existing link a second time.

By explicitly using unique links in the guide and manual,
this race condition is avoided. The final paths to the
documents kde looks are still the same as before this change.
The only difference is that the symlink has been shifted
one level down in the full path.