cmake AddHtmlTarget: use config_file() for copying files
Rather than making custom commands to simply copy files in place we can
depend on cmake' configure_file() command to do the right thing for us.
That is configure_file() will track any file changes for us and
automatically copies the updated files to where we told it initially.

This change requires an explicit rerun of cmake!